move getPauseReason to utils
